chore: Migrate websocket handling to socket.io
