+++ backend/package.json
@@ -1,42 +1,39 @@
 {
-  "name": "quizz-backend",
-  "version": "1.0.0",
-  "description": "A real-time multiplayer quiz application.",
-  "main": "src/index.ts",
-  "scripts": {
-    "start": "ts-node src/index.ts",
-    "watch": "nodemon src/index.ts",
-    "build": "tsc --build"
-  },
-  "dependencies": {
-    "axios": "^0.27.2",
-    "cors": "^2.8.5",
-    "discord.js": "^13.8.1",
-    "dotenv": "^15.0.1",
-    "express": "^4.17.2",
-    "express-ws": "^5.0.2",
-    "joi": "^17.6.0",
-    "mongodb": "^4.3.1",
-    "redis": "^4.0.3",
-    "reflect-metadata": "^0.1.13",
-    "ws": "^8.11.0"
-  },
-  "devDependencies": {
-    "@types/cors": "^2.8.12",
-    "@types/express": "^4.17.13",
-    "@types/node": "^18.6.4",
-    "nodemon": "^2.0.19",
-    "prettier": "^2.5.1",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.7.4"
-  },
-  "prettier": {
-    "printWidth": 250,
-    "singleQuote": true,
-    "semi": true,
-    "useTabs": false,
-    "tabWidth": 4,
-    "trailingComma": "none",
-    "arrowParens": "always"
-  }
+    "name": "quizz-backend",
+    "version": "1.0.0",
+    "description": "A real-time multiplayer quiz application.",
+    "main": "src/index.ts",
+    "scripts": {
+        "start": "ts-node src/index.ts",
+        "watch": "nodemon src/index.ts",
+        "build": "tsc --build"
+    },
+    "dependencies": {
+        "axios": "^0.27.2",
+        "cors": "^2.8.5",
+        "dotenv": "^15.0.1",
+        "express": "^4.17.2",
+        "express-ws": "^5.0.2",
+        "joi": "^17.6.0",
+        "reflect-metadata": "^0.1.13",
+        "socket.io": "^4.7.2"
+    },
+    "devDependencies": {
+        "@types/cors": "^2.8.12",
+        "@types/express": "^4.17.13",
+        "@types/node": "^18.6.4",
+        "nodemon": "^2.0.19",
+        "prettier": "^2.5.1",
+        "ts-node": "^10.9.1",
+        "typescript": "^4.7.4"
+    },
+    "prettier": {
+        "printWidth": 250,
+        "singleQuote": true,
+        "semi": true,
+        "useTabs": false,
+        "tabWidth": 4,
+        "trailingComma": "none",
+        "arrowParens": "always"
+    }
 }
