+++ frontend/package.json
@@ -1,52 +1,53 @@
 {
-  "name": "quizz",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@fontsource/inter": "^4.5.14",
-    "@fontsource/lexend-deca": "^4.5.12",
-    "@fontsource/mulish": "^4.5.13",
-    "@fontsource/poppins": "^4.5.10",
-    "@fontsource/source-sans-pro": "^4.5.11",
-    "@reduxjs/toolkit": "^1.9.0",
-    "@types/react-redux": "^7.1.24",
-    "classnames": "^2.3.2",
-    "clone": "^2.1.2",
-    "framer-motion": "^8.0.2",
-    "install": "^0.13.0",
-    "next": "13.0.0",
-    "npm": "^9.2.0",
-    "react": "18.0.0",
-    "react-dom": "18.0.0",
-    "react-hook-form": "^7.39.4",
-    "react-icons": "^4.6.0",
-    "react-redux": "^8.0.5"
-  },
-  "devDependencies": {
-    "@testing-library/react": "^13.4.0",
-    "@types/clone": "^2.1.1",
-    "@types/node": "17.0.23",
-    "@types/react": "17.0.43",
-    "@types/react-dom": "17.0.14",
-    "eslint": "8.12.0",
-    "eslint-config-next": "12.1.4",
-    "sass": "^1.39.0",
-    "scss": "^0.2.4",
-    "typescript": "4.6.3"
-  },
-  "prettier": {
-    "printWidth": 250,
-    "singleQuote": true,
-    "semi": true,
-    "useTabs": false,
-    "tabWidth": 4,
-    "trailingComma": "none",
-    "arrowParens": "always"
-  }
+    "name": "quizz",
+    "version": "0.1.0",
+    "private": true,
+    "scripts": {
+        "dev": "next dev",
+        "build": "next build",
+        "start": "next start",
+        "lint": "next lint"
+    },
+    "dependencies": {
+        "@fontsource/inter": "^4.5.14",
+        "@fontsource/lexend-deca": "^4.5.12",
+        "@fontsource/mulish": "^4.5.13",
+        "@fontsource/poppins": "^4.5.10",
+        "@fontsource/source-sans-pro": "^4.5.11",
+        "@reduxjs/toolkit": "^1.9.0",
+        "@types/react-redux": "^7.1.24",
+        "classnames": "^2.3.2",
+        "clone": "^2.1.2",
+        "framer-motion": "^8.0.2",
+        "install": "^0.13.0",
+        "next": "13.0.0",
+        "npm": "^9.2.0",
+        "react": "18.0.0",
+        "react-dom": "18.0.0",
+        "react-hook-form": "^7.39.4",
+        "react-icons": "^4.6.0",
+        "react-redux": "^8.0.5",
+        "socket.io-client": "^4.7.2"
+    },
+    "devDependencies": {
+        "@testing-library/react": "^13.4.0",
+        "@types/clone": "^2.1.1",
+        "@types/node": "17.0.23",
+        "@types/react": "17.0.43",
+        "@types/react-dom": "17.0.14",
+        "eslint": "8.12.0",
+        "eslint-config-next": "12.1.4",
+        "sass": "^1.39.0",
+        "scss": "^0.2.4",
+        "typescript": "4.6.3"
+    },
+    "prettier": {
+        "printWidth": 250,
+        "singleQuote": true,
+        "semi": true,
+        "useTabs": false,
+        "tabWidth": 4,
+        "trailingComma": "none",
+        "arrowParens": "always"
+    }
 }
